requirements.txt
ffmpeg-python
onvif-zeep
opencv-python
pandas
pillow
